+++ weight_stream/__init__.py
@@ -8,7 +8,7 @@ Usage:
     output = model.generate("Hello, world")
 """
 
-__version__ = "0.11.0"
+__version__ = "0.12.0"
 
 from .backends.llama_cpp import WeightStreamModel
 from .core.exceptions import (